Add hydro and set command extraction
