+++ tests/integration/test_integration_variantmanager_tool.py
@@ -51,14 +51,9 @@ def generate_study(
 
 def test_variant_manager(app: FastAPI):
     client = TestClient(app, raise_server_exceptions=False)
-    res = client.post(
-        "/v1/login", json={"username": "admin", "password": "admin"}
+    commands = CLIVariantManager.parse_commands(
+        test_dir / "assets" / "commands1.json"
     )
-    admin_credentials = res.json()
-    vm = CLIVariantManager(
-        session=client, token=admin_credentials["access_token"]
-    )
-    commands = vm.parse_commands(test_dir / "assets" / "commands1.json")
     res = generate_study(client, "test", 720, commands)
     assert res is not None and res.success
 
@@ -70,32 +65,31 @@ def test_parse_commands(tmp_path: str, app: FastAPI):
     # todo add a quite feature-exhaustive test study
     base_dir = Path("/home/buiquangpau/scratch/test_antares_vm")
     export_path = Path(tmp_path) / "commands"
-    for study in os.listdir(base_dir):
-        if "038" not in study:
-            continue
-        study_path = base_dir / study
-        output_dir = Path(export_path) / study
-        logger.info(study_path)
-        logger.info(output_dir)
-        try:
-            study_info = IniReader().read(study_path / "study.antares")
-            version = study_info["antares"]["version"]
-            name = study_info["antares"]["caption"]
-            client = TestClient(app, raise_server_exceptions=False)
-            res = client.post(
-                "/v1/login", json={"username": "admin", "password": "admin"}
-            )
-            admin_credentials = res.json()
-            vm = CLIVariantManager(
-                session=client, token=admin_credentials["access_token"]
-            )
+    # for study in os.listdir(base_dir):
+    #     if "038" not in study:
+    #         continue
+    #     study_path = base_dir / study
+    study = "Zonal_CH_BAU_2025_LF"
+    study_path = Path(
+        "/home/buiquangpau/scratch/antares_workspace_tmp/Zonal_CH_BAU_2025_LF"
+    )
+    output_dir = Path(export_path) / study
+    logger.info(study_path)
+    logger.info(output_dir)
+    try:
+        study_info = IniReader().read(study_path / "study.antares")
+        version = study_info["antares"]["version"]
+        name = study_info["antares"]["caption"]
+        client = TestClient(app, raise_server_exceptions=False)
 
-            CLIVariantManager.extract_commands(study_path, output_dir)
-            commands = vm.parse_commands(output_dir / "commands.json")
-            res = generate_study(
-                client, name, version, commands, output_dir / "matrices"
-            )
-            logger.info(res.json())
-            #        assert res is not None and res.success
-        except Exception as e:
-            logger.error(f"Failure on {study_path}", exc_info=e)
+        CLIVariantManager.extract_commands(study_path, output_dir)
+        commands = CLIVariantManager.parse_commands(
+            output_dir / "commands.json"
+        )
+        res = generate_study(
+            client, name, version, commands, output_dir / "matrices"
+        )
+        logger.info(res.json())
+        #        assert res is not None and res.success
+    except Exception as e:
+        logger.error(f"Failure on {study_path}", exc_info=e)
